+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Unreleased
 .. vendor-insert-here
 
 - Update vendored schemas (2022-11-14)
+- Downloads of schemas from remote (http/https) locations will now retry if the
+  downloaded data is not valid JSON (:issue:`183`)
 
 0.19.1
 ------