+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Unreleased
 
 .. vendor-insert-here
 
+0.19.2
+------
+
 - Update vendored schemas (2022-11-14)
 - Downloads of schemas from remote (http/https) locations will now retry if the
   downloaded data is not valid JSON (:issue:`183`)